voicebot_with_avatar/requirements.txt
python-dotenv
livekit
cartesia
livekit-agents
flask
flask-cors
